CI: add new wolfi-vm repo identity for chainctl

+++ Makefile
@@ -33,19 +33,22 @@ manifest:
 		"cgr.dev/chainguard-private/${DOCKER_IMAGE}"  | jq -r .payload | base64 -d | jq .predicate > tmp-manifest.json || rm -f tmp-manifest.json
 
 initrd:
+	[ -z "${HTTP_AUTH}" ] && HTTP_AUTH="basic:apk.cgr.dev:user:$(shell chainctl auth token --audience apk.cgr.dev)"
 	ls tmp-manifest.json || $(MAKE) manifest
 	# same as .mkinird but we pass a manifest and a cloud provider to behave accordingly
 	ls initrd*gz 2>/dev/null || docker run --rm -i \
-		-e HTTP_AUTH="basic:apk.cgr.dev:user:$(shell chainctl auth token --audience apk.cgr.dev)" \
+		-e HTTP_AUTH="${HTTP_AUTH}" \
 		--mount type=bind,source="$(TOP_D)",destination="/srv" \
 		cgr.dev/chainguard/wolfi-base:latest \
 		/srv/.mkinitrd $(shell id -u):$(shell id -g) $(shell uname -m) "/srv/tmp-manifest.json"
 
 disk: initrd
 	@if ! ls disk*${DOCKER_IMAGE}* 2>/dev/null; then\
+		[ -z "${HTTP_AUTH}" ] && HTTP_AUTH="basic:apk.cgr.dev:user:$(shell chainctl auth token --audience apk.cgr.dev)"
 		docker run --privileged --rm -i \
 			-e DOCKER_IMAGE=${DOCKER_IMAGE} \
-			-e HTTP_AUTH="basic:apk.cgr.dev:user:$(shell chainctl auth token --audience apk.cgr.dev)" \
+			-e HTTP_AUTH="${HTTP_AUTH}" \
+		--mount type=bind,source="$(TOP_D)",destination="/srv" \
 			--mount type=bind,source="$(TOP_D)",destination="/srv" \
 			cgr.dev/chainguard/wolfi-base:latest \
 			/srv/.mkdiskimg $(shell id -u):$(shell id -g) /srv/$(shell ls vmlinuz* | sort -V | tail -n1) /srv/$(shell ls initrd* | sort -V | tail -n1);\
